Sample view on double click (#153)
Add clicklistener for doubleclicks on the project overview
