+++ sample-tracking-status-overview-app/src/main/groovy/life/qbic/portal/sampletracking/components/AppView.groovy
@@ -1,11 +1,16 @@
 package life.qbic.portal.sampletracking.components
 
+import com.vaadin.ui.Button
 import com.vaadin.ui.Component
+import com.vaadin.ui.Grid
 import com.vaadin.ui.HorizontalLayout
 import com.vaadin.ui.Label
 import com.vaadin.ui.VerticalLayout
+import com.vaadin.ui.components.grid.ItemClickListener
 import com.vaadin.ui.themes.ValoTheme
 import life.qbic.portal.sampletracking.components.projectoverview.ProjectOverviewView
+import life.qbic.portal.sampletracking.components.projectoverview.ProjectSummary
+import life.qbic.portal.sampletracking.components.sampleoverview.SampleOverviewController
 import life.qbic.portal.sampletracking.components.sampleoverview.SampleOverviewView
 import life.qbic.portal.sampletracking.components.toggle.ToggleButton
 
@@ -20,11 +25,13 @@ class AppView extends VerticalLayout {
 
     private final ProjectOverviewView projectOverviewView
     private final SampleOverviewView sampleOverviewView
+    private final SampleOverviewController sampleOverviewController
+
     private final HorizontalLayout hotbar = new HorizontalLayout()
     protected final Label titleLabel = new Label("Sample Status Portlet")
     private final ToggleButton projectSampleToggle
 
-    AppView(ProjectOverviewView projectOverviewView, SampleOverviewView sampleOverviewView) {
+    AppView(ProjectOverviewView projectOverviewView, SampleOverviewView sampleOverviewView, SampleOverviewController sampleOverviewController) {
         this.setMargin(true)
         this.setSpacing(true)
 
@@ -32,6 +39,7 @@ class AppView extends VerticalLayout {
 
         this.projectOverviewView = projectOverviewView
         this.sampleOverviewView = sampleOverviewView
+        this.sampleOverviewController = sampleOverviewController
 
         projectSampleToggle = setupProjectSampleToggle()
 
@@ -47,11 +55,12 @@ class AppView extends VerticalLayout {
         this.addComponents(titleLabel, createSpacer(2, Unit.EM), hotbar, projectOverviewView, sampleOverviewView)
     }
 
+
     private void listenToProjectSelectionChange(){
         projectOverviewView.onSelectedProjectChange({
             if (it) {
                 projectSampleToggle.setEnabled(true)
-                sampleOverviewView.loadProjectSamples(it.code)
+                sampleOverviewController.getSamplesFor(it.code)
             } else {
                 projectSampleToggle.setEnabled(false)
                 sampleOverviewView.reset()
@@ -87,6 +96,11 @@ class AppView extends VerticalLayout {
         }
         projectSampleToggle.addClickListener(switchToSampleView, ToggleButton.State.ONE)
         projectSampleToggle.addClickListener(switchToProjectView, ToggleButton.State.TWO)
+
+        projectOverviewView.onProjectDoubleClick({
+            sampleOverviewController.getSamplesFor(it.code)
+            projectSampleToggle.click()
+        })
     }
 
     private static Component createSpacer(float height, Unit unit) {
