+++ sample-tracking-status-overview-app/src/main/groovy/life/qbic/portal/sampletracking/components/sampleoverview/SampleOverviewView.groovy
@@ -1,13 +1,16 @@
 package life.qbic.portal.sampletracking.components.sampleoverview
 
-import com.vaadin.ui.Button
+import com.vaadin.data.provider.ListDataProvider
+import com.vaadin.shared.ui.grid.HeightMode
+import com.vaadin.ui.Grid
 import com.vaadin.ui.HorizontalLayout
 import com.vaadin.ui.VerticalLayout
+import life.qbic.business.samples.Sample
+import life.qbic.business.samples.info.GetSamplesInfoOutput
+import life.qbic.portal.sampletracking.communication.notification.NotificationService
 import life.qbic.portal.sampletracking.components.HasHotbar
 import life.qbic.portal.sampletracking.components.HasTitle
 import life.qbic.portal.sampletracking.components.Resettable
-import life.qbic.portal.sampletracking.components.sampleoverview.samplelist.ProjectSamplesController
-import life.qbic.portal.sampletracking.components.sampleoverview.samplelist.ProjectSamplesView
 
 /**
  * <b>A view component showing a list of samples</b>
@@ -18,24 +21,72 @@ class SampleOverviewView extends VerticalLayout implements HasHotbar, HasTitle, 
 
     private static final String TITLE = "Sample Overview"
 
-    private final ProjectSamplesController projectSamplesController
-    private final ProjectSamplesView projectSamplesView
+    private final ViewModel viewModel
+    private final Presenter presenter
+    private Grid<Sample> samplesGrid
+    private final HorizontalLayout hotbar = new HorizontalLayout()
 
-    private final HorizontalLayout hotbar = new HorizontalLayout(new Button("Sample Klick"))
+    SampleOverviewView(NotificationService notificationService) {
+        this.viewModel = new ViewModel()
+        this.presenter = new Presenter(notificationService, viewModel)
 
-    SampleOverviewView(ProjectSamplesController projectSamplesController, ProjectSamplesView projectSamplesView) {
-        this.projectSamplesController = projectSamplesController
-        this.projectSamplesView = projectSamplesView
-        this.addComponents(this.projectSamplesView)
+        initLayout()
+    }
+
+    private void initLayout() {
         this.setMargin(false)
+        this.setSizeFull()
+
+        this.samplesGrid = createSamplesGrid(viewModel.samples)
+        samplesGrid.setSizeFull()
+        this.addComponents(samplesGrid)
+    }
+
+    Presenter getPresenter() {
+        return presenter
+    }
+
+    private static Grid<Sample> createSamplesGrid(Collection<Sample> samples) {
+        Grid<Sample> samplesGrid = new Grid<>()
+        ListDataProvider<Sample> dataProvider = ListDataProvider.ofCollection(samples)
+        samplesGrid.addColumn(Sample::getCode).setCaption("Sample Code").setId("SampleCode")
+        samplesGrid.addColumn(Sample::getName).setCaption("Sample Name").setId("SampleName")
+        samplesGrid.addColumn(Sample::getStatus).setCaption("Sample Status").setId("SampleStatus")
+        samplesGrid.setSelectionMode(Grid.SelectionMode.NONE)
+        samplesGrid.setDataProvider(dataProvider)
+        samplesGrid.setHeightMode(HeightMode.ROW)
+        return samplesGrid
     }
 
     /**
-     * Loads the project samples for a given project code
-     * @param projectCode The code for which samples are loaded
+     * Model for this view component
      */
-    void loadProjectSamples(String projectCode){
-        projectSamplesController.getSamplesFor(projectCode)
+    private static class ViewModel {
+        List<Sample> samples = []
+    }
+
+    /**
+     * Presenter filling the grid model with information
+     */
+    private static class Presenter implements GetSamplesInfoOutput {
+        private final NotificationService notificationService
+        private final ViewModel viewModel
+
+        Presenter(NotificationService notificationService, ViewModel viewModel) {
+            this.notificationService = notificationService
+            this.viewModel = viewModel
+        }
+
+        @Override
+        void failedExecution(String reason) {
+            notificationService.publishFailure("Could not load samples: $reason")
+        }
+
+        @Override
+        void samplesWithNames(Collection<Sample> samples) {
+            viewModel.samples.clear()
+            viewModel.samples.addAll(samples)
+        }
     }
 
     @Override
@@ -50,6 +101,7 @@ class SampleOverviewView extends VerticalLayout implements HasHotbar, HasTitle, 
 
     @Override
     void reset() {
-        projectSamplesView.reset()
+        viewModel.samples.clear()
     }
+
 }
